Changing the $watchCollection function in controller.js to $watch with extra argument True
Once the array elements passed to the chart-data changed, the pie chart
is not reflecting the updated data. This is because of use of
$watchCollection which does not check the value of the Array element, it
checks only its size. so we need to move back to $watch with the array
and using an extra Argument TRUE, that will automatically scan the
changes and will refresh the Chart once any of the array item value
changes or the size of the array changes.
Reference:
http://teropa.info/blog/2014/01/26/the-three-watch-depths-of-angularjs.html